1->Weather and Forecast Mapper fixed. 2->Display current weather 3->Weather icons url fixed.

+++ app/src/main/java/com/hashim/mapswithgeofencing/network/response/weather/Main.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2021/  3/ 21.  Created by Hashim Tahir
+ * Copyright (c) 2021/  3/ 25.  Created by Hashim Tahir
  */
 
 package com.hashim.mapswithgeofencing.network.response.weather
@@ -17,7 +17,7 @@ data class Main(
         @SerializedName("temp")
         val temp: Double,
         @SerializedName("temp_max")
-        val tempMax: Int,
+        val tempMax: Double,
         @SerializedName("temp_min")
         val tempMin: Double
 )
